pyproject.toml
[project]
name = "bad-color-map"
version = "0.1.0"
description = ""
authors = [
    { name = "Thomas Holland", email = "101332229+0thomasholland@users.noreply.github.com" },
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = ["numpy (>=2.3.5,<3.0.0)", "pillow (>=12.0.0,<13.0.0)", "matplotlib (>=3.10.7,<4.0.0)", "importlib (>=1.0.4,<2.0.0)", "colorspacious (>=1.1.2,<2.0.0)"]

[tool.poetry]
packages = [{ include = "bad_color_map", from = "src" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
